backend/notify-main/requirements.txt
flask
flask_sqlalchemy
flask_mail
psycopg2-binary
python-dotenv
